[OMNI-4937] Integration with Online Payment with Gateway

+++ src/Omni/Observer/OrderObserver.php
@@ -1,4 +1,5 @@
 <?php
+
 namespace Ls\Omni\Observer;
 
 use \Ls\Core\Model\LSR;
@@ -14,16 +15,16 @@ use Magento\Framework\Event\ObserverInterface;
  */
 class OrderObserver implements ObserverInterface
 {
-    /** @var ContactHelper  */
+    /** @var ContactHelper */
     private $contactHelper;
 
-    /** @var BasketHelper  */
+    /** @var BasketHelper */
     private $basketHelper;
 
-    /** @var OrderHelper  */
+    /** @var OrderHelper */
     private $orderHelper;
 
-    /** @var \Psr\Log\LoggerInterface  */
+    /** @var \Psr\Log\LoggerInterface */
     private $logger;
 
     /** @var \Magento\Customer\Model\Session\Proxy $customerSession */
@@ -32,13 +33,13 @@ class OrderObserver implements ObserverInterface
     /** @var \Magento\Checkout\Model\Session\Proxy $checkoutSession */
     private $checkoutSession;
 
-    /** @var bool  */
+    /** @var bool */
     private $watchNextSave = false;
 
     /** @var \Magento\Sales\Model\ResourceModel\Order $orderResourceModel */
     private $orderResourceModel;
 
-    /** @var \Ls\Core\Model\LSR @var  */
+    /** @var \Ls\Core\Model\LSR @var */
     private $lsr;
 
     /**
@@ -70,13 +71,14 @@ class OrderObserver implements ObserverInterface
         $this->customerSession = $customerSession;
         $this->checkoutSession = $checkoutSession;
         $this->orderResourceModel = $orderResourceModel;
-        $this->lsr  =   $LSR;
+        $this->lsr = $LSR;
     }
 
     /**
      * @param \Magento\Framework\Event\Observer $observer
      * @return $this|void
      * @throws \Ls\Omni\Exception\InvalidEnumException
+     * @throws \Magento\Framework\Exception\InputException
      * @throws \Magento\Framework\Exception\NoSuchEntityException
      */
     public function execute(\Magento\Framework\Event\Observer $observer)
@@ -85,37 +87,55 @@ class OrderObserver implements ObserverInterface
          * Adding condition to only process if LSR is enabled.
          */
         if ($this->lsr->isLSR()) {
+            $success = false;
             $order = $observer->getEvent()->getData('order');
-            /** @var Entity\Order $oneListCalculation */
             $oneListCalculation = $this->basketHelper->getOneListCalculation();
-            $request = $this->orderHelper->prepareOrder($order, $oneListCalculation);
-            $response = $this->orderHelper->placeOrder($request);
-            try {
-                if ($response) {
-                    //delete from Omni.
-                    $documentId = $response->getDocumentId();
-                    $order->setDocumentId($documentId);
-                    $this->orderResourceModel->save($order);
-                    $this->checkoutSession->setLastDocumentId($documentId);
-                    $this->checkoutSession->unsetData('member_points');
-                    if ($this->customerSession->getData(LSR::SESSION_CART_ONELIST)) {
-                        $onelist = $this->customerSession->getData(LSR::SESSION_CART_ONELIST);
-                        //TODO error which Hjalti highlighted. when there is only one item in the cart and customer remove that.
-                        $success = $this->basketHelper->delete($onelist);
-                        $this->customerSession->unsetData(LSR::SESSION_CART_ONELIST);
-                        // delete checkout session data.
-                        $this->basketHelper->unSetOneListCalculation();
-                    }
-                } else {
-                    // TODO: error handling
-                    $this->logger->critical(
-                        __('Something trrible happen while placing order')
-                    );
-                }
-            } catch (\Exception $e) {
-                $this->logger->error($e->getMessage());
+            if (empty($order)) {
+                $orderIds = $observer->getEvent()->getOrderIds();
+                $order = $this->orderHelper->orderRepository->get($orderIds[0]);
             }
+            //checking for Adyen payment gateway
+            $adyen_response = $observer->getEvent()->getData('adyen_response');
+            if (!empty($adyen_response)) {
+                $order->getPayment()->setLastTransId($adyen_response['pspReference']);
+                $order->getPayment()->setCcTransId($adyen_response['pspReference']);
+                $order->getPayment()->setCcType($adyen_response['paymentMethod']);
+                $order->getPayment()->setCcStatus($adyen_response['authResult']);
+                $this->orderHelper->orderRepository->save($order);
+                $order = $this->orderHelper->orderRepository->get($order->getEntityId());
+            }
+            $paymentMethod = $order->getPayment()->getMethodInstance();
+            $transId = $order->getPayment()->getLastTransId();
+            if (($paymentMethod->isOffline() == true || !empty($transId)) && !empty($oneListCalculation)) {
+                $request = $this->orderHelper->prepareOrder($order, $oneListCalculation);
+                $response = $this->orderHelper->placeOrder($request);
+                try {
+                    if ($response) {
+                        //delete from Omni.
+                        $documentId = $response->getDocumentId();
+                        $order->setDocumentId($documentId);
+                        $this->orderResourceModel->save($order);
+                        $this->checkoutSession->setLastDocumentId($documentId);
+                        $this->checkoutSession->unsetData('member_points');
+                        if ($this->customerSession->getData(LSR::SESSION_CART_ONELIST)) {
+                            $onelist = $this->customerSession->getData(LSR::SESSION_CART_ONELIST);
+                            //TODO error which Hjalti highlighted. when there is only one item in the cart and customer remove that.
+                            $success = $this->basketHelper->delete($onelist);
+                            $this->customerSession->unsetData(LSR::SESSION_CART_ONELIST);
+                            // delete checkout session data.
+                            $this->basketHelper->unSetOneListCalculation();
+                        }
+                    } else {
+                        // TODO: error handling
+                        $this->logger->critical(
+                            __('Something trrible happen while placing order')
+                        );
+                    }
+                } catch (\Exception $e) {
+                    $this->logger->error($e->getMessage());
+                }
+            }
+            return $this;
         }
-        return $this;
     }
 }
